Basic project structure
This commit contains basic structure for the project. This first aim covers the pre-processing that we need to perform to Hermes raw data in order to convert it to non aggregated data. This particular commit contains autogenerated code for that part that hasn't been tested and may not work as expected.